Add files via upload
Task 1: To-Do List Application
Description:
A To-Do List application assists users in organizing and managing their tasks efficiently. This project involves creating a command-line or GUI-based application in Python, enabling users to:

1.Create: Users can add tasks to their to-do lists.
2.Update: Modify existing tasks or their details.
3.Track: Keep a record of tasks, including their status (completed or pending).
4.Delete: Remove tasks from the list when they are no longer needed.
5.Display: View the list of tasks to see what needs to be done.

Task 2: Simple Calculator
Description:
The Simple Calculator project involves building a program that performs basic arithmetic operations. The user is prompted to input two numbers and select an operation (addition, subtraction, multiplication, division). The calculator then computes the result and displays it to the user.

Task 3: Password Generator
Description:
A Password Generator application creates strong and random passwords based on user-specified criteria. Users can determine the length and complexity of the password they want. The application generates a password using a combination of random characters and displays it to the user.

Task 4: Rock-Paper-Scissors Game
Description:
The Rock-Paper-Scissors Game involves building a program that allows users to play the classic game against the computer. The user chooses rock, paper, or scissors, and the computer randomly selects an option. The program then determines the winner based on the game's rules and displays the outcome.

Task 5: Contact Book
Description:
A Contact Book application stores and manages contact information. Users can perform various operations such as:

1.Add Contact: Enter new contacts with details like name, phone number, email, and address.
2.View Contact List: Display a list of all saved contacts with names and phone numbers.
3.Search Contact: Implement a search function to find contacts by name or phone number.
4.Update Contact: Modify or update existing contact details.
5.Delete Contact: Remove contacts from the list.